# docs/swagger.yaml
basePath: /book
definitions:
...
  model.PoolState:
    enum:
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - PoolStateAdd
    - PoolStateRemove
    - PoolStateUsed
    - PoolStateClaim
  model.PoolType:
    enum:
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - PoolTypeTick
    - PoolTypeBtc
    - PoolTypeBoth
...
  request.OrderPoolBrc20UpdateReq:
    properties:
      net:
        description: livenet/signet/testnet
        type: string
      orderId:
        type: string
      poolState:
        allOf:
        - $ref: '#/definitions/model.PoolState'
        description: 1-add,2-remove,3-used,4-claim
    type: object
  request.PoolBrc20ClaimReq:
    properties:
      address:
        type: string
      net:
        description: livenet/signet/testnet
        type: string
      poolOrderId:
        type: string
      preSigScript:
        type: string
      tick:
        type: string
    type: object
  request.PoolBrc20ClaimUpdateReq:
    properties:
      poolOrderId:
        type: string
      psbtRaw:
        type: string
    type: object
  request.PoolBrc20PushReq:
    properties:
      address:
        type: string
      amount:
        type: integer
      coinAmount:
        type: integer
      coinPsbtRaw:
        type: string
      net:
        description: livenet/signet/testnet
        type: string
      pair:
        type: string
      poolState:
        allOf:
        - $ref: '#/definitions/model.PoolState'
        description: 1-add,2-remove,3-used,4-claim
      poolType:
        allOf:
        - $ref: '#/definitions/model.PoolType'
        description: 1-tick,2-btc,3-both
      psbtRaw:
        type: string
      tick:
        type: string
    type: object
...
  respond.Message:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      processingTime:
        type: integer
    type: object
...
  respond.PoolBrc20ClaimResp:
    properties:
      coinAmount:
        description: Brc20 amount
        type: integer
      coinPsbtRaw:
        description: coin PSBT Raw
        type: string
      fee:
        description: claim fee
        type: integer
      inscriptionId:
        description: InscriptionId
        type: string
      net:
        description: Net env
        type: string
      orderId:
        description: Order ID
        type: string
      psbtRaw:
        description: PSBT Raw
        type: string
      rewardCoinAmount:
        type: integer
      tick:
        description: Brc20 symbol
        type: string
    type: object
  respond.PoolBrc20Item:
    properties:
      address:
        description: address
        type: string
      amount:
        type: integer
      coinAddress:
        description: tick
        type: string
      coinAmount:
        description: tick
        type: integer
      coinDecimalNum:
        description: tick
        type: integer
      coinPsbtRaw:
        description: coin PSBT Raw
        type: string
      decimalNum:
        type: integer
      inscriptionId:
        description: InscriptionId
        type: string
      net:
        description: Net env
        type: string
      orderId:
        description: Order ID
        type: string
      pair:
        description: Brc20 pair
        type: string
      poolState:
        allOf:
        - $ref: '#/definitions/model.PoolState'
        description: pool state：1-add,2-remove,3-used,4-claim
      poolType:
        allOf:
        - $ref: '#/definitions/model.PoolType'
        description: pool type：1-tick,2-btc
      psbtRaw:
        description: PSBT Raw
        type: string
      tick:
        description: Brc20 symbol
        type: string
      timestamp:
        description: Create time
        type: integer
    type: object
  respond.PoolInfoItem:
    properties:
      amount:
        description: 'Btc: sat'
        type: integer
      coinAmount:
        type: integer
      coinDecimalNum:
        description: omitempty
        type: integer
      decimalNum:
        description: Btc decimal
        type: integer
      net:
        description: Net env
        type: string
      ownAmount:
        type: integer
      ownCoinAmount:
        type: integer
      ownCount:
        type: integer
      pair:
        description: Brc20 pair
        type: string
      tick:
        description: Brc20 symbol
        type: string
    type: object
  respond.PoolInfoResponse:
    properties:
      flag:
        type: integer
      results:
        items:
          $ref: '#/definitions/respond.PoolInfoItem'
        type: array
      total:
        type: integer
    type: object
  respond.PoolInscriptionItem:
    properties:
      coinAmount:
        type: string
      inscriptionId:
        type: string
      inscriptionNumber:
        type: string
    type: object
  respond.PoolInscriptionResp:
    properties:
      availableList:
        items:
          $ref: '#/definitions/respond.PoolInscriptionItem'
        type: array
      net:
        type: string
      tick:
        type: string
      total:
        type: integer
    type: object
  respond.PoolKeyInfoResp:
    properties:
      net:
        description: Net env
        type: string
      publicKey:
        description: key
        type: string
    type: object
  respond.PoolResponse:
    properties:
      flag:
        type: integer
      results:
        items:
          $ref: '#/definitions/respond.PoolBrc20Item'
        type: array
      total:
        type: integer
    type: object
...
info:
  contact:
    name: API Support
  description: OrdBook API Service
  title: OrdBook API Service
  version: "1.0"
paths:
...
  /brc20/pool/inscription:
    get:
      description: Fetch pool inscription
      parameters:
      - description: net
        in: query
        name: net
        type: string
      - description: tick
        in: query
        name: tick
        type: string
      - description: address
        in: query
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.PoolInscriptionResp'
      summary: Fetch pool inscription
      tags:
      - brc20
  /brc20/pool/order/{orderId}:
    get:
      description: Fetch one pool order
      parameters:
      - description: orderId
        in: path
        name: orderId
        required: true
        type: string
      - description: address
        in: query
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.PoolBrc20Item'
      summary: Fetch one pool order
      tags:
      - brc20
  /brc20/pool/order/claim:
    post:
      description: get claim pool order
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.PoolBrc20ClaimReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.PoolBrc20ClaimResp'
      summary: get claim pool order
      tags:
      - brc20
  /brc20/pool/order/claim/commit:
    post:
      description: claim pool order
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.PoolBrc20ClaimUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Message'
      summary: claim pool order
      tags:
      - brc20
  /brc20/pool/order/push:
    post:
      description: Push pool order
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.PoolBrc20PushReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Message'
      summary: Push pool order
      tags:
      - brc20
  /brc20/pool/order/update:
    post:
      description: Update pool order
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.OrderPoolBrc20UpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Message'
      summary: Update pool order
      tags:
      - brc20
  /brc20/pool/orders:
    get:
      description: Fetch pool orders
      parameters:
      - description: net:mainnet/signet/testnet
        in: query
        name: net
        type: string
      - description: tick
        in: query
        name: tick
        type: string
      - description: address
        in: query
        name: address
        type: string
      - description: 'poolState: 1-add,2-remove,3-used,4-claim'
        in: query
        name: poolState
        type: integer
      - description: 'poolType: 1-tick,2-btc'
        in: query
        name: poolType
        type: integer
      - description: 'limit: Max-50'
        in: query
        name: limit
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: flag
        in: query
        name: flag
        type: integer
      - description: 'sortKey: timestamp/coinRatePrice, default:timestamp'
        in: query
        name: sortKey
        type: string
      - description: 'sortType: 1/-1'
        in: query
        name: sortType
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.PoolResponse'
      summary: Fetch pool orders
      tags:
      - brc20
  /brc20/pool/pair/info:
    get:
      description: Fetch pool pair info
      parameters:
      - description: net
        in: query
        name: net
        type: string
      - description: tick
        in: query
        name: tick
        type: string
      - description: pair
        in: query
        name: pair
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.PoolInfoResponse'
      summary: Fetch pool pair info
      tags:
      - brc20
  /brc20/pool/pair/info/one:
    get:
      description: Fetch one pool pair info
      parameters:
      - description: net
        in: query
        name: net
        type: string
      - description: tick
        in: query
        name: tick
        type: string
      - description: pair
        in: query
        name: pair
        type: string
      - description: address
        in: query
        name: address
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.PoolInfoItem'
      summary: Fetch one pool pair info
      tags:
      - brc20
  /brc20/pool/pair/key:
    get:
      description: Fetch pool key
      parameters:
      - description: net
        in: query
        name: net
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.PoolKeyInfoResp'
      summary: Fetch pool key
      tags:
      - brc20
...
schemes:
- https
swagger: "2.0"